requirements.txt
asgiref==3.7.2
astroid==3.0.2
dill==0.3.7
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
Faker==22.1.0
isort==5.13.2
lazy-object-proxy==1.10.0
mccabe==0.7.0
platformdirs==4.1.0
pylint==3.0.3
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
typed-ast==1.5.5
typing_extensions==4.9.0
validate-docbr==1.10.0
wrapt==1.16.0
